api/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Search
meilisearch-python-async==1.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI & Automation
openai==1.10.0
langchain==0.1.0
httpx==0.26.0
aiofiles==23.2.1

# Payments
stripe==7.10.0
mercadopago==2.2.1

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
Pillow==10.2.0
python-magic==0.4.27
boto3==1.34.20  # For S3 storage

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
